:wrench: Fix minor bug

+++ src/main/kotlin/com/belveth/bullettime/domain/user/repository/UserRepository.kt
@@ -4,6 +4,6 @@ import com.belveth.bullettime.domain.user.entity.UserEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.Optional
 
-interface UserRepository: JpaRepository<UserEntity, Long> {
+interface UserRepository: JpaRepository<UserEntity, Int> {
   fun getUser(id: Int): Optional<UserEntity>
 }
